+++ src/js/attendee.client.js
@@ -1,7 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { Router, Route, useRouterHistory, IndexRoute, IndexLink } from "react-router";
-import { createHashHistory } from 'history'
 
 require('!style!css!sass!../sass/main.scss');
 
@@ -15,25 +14,22 @@ import Profile from "./pages/Profile";
 import Info from "./pages/Info";
 import Login from "./pages/Login";
 
+import AppHistory from './util/AppHistory';
 import AuthService from './util/AuthService';
 import * as config from './auth0.config.js';
 
 const app = document.getElementById('app');
-//If we find that we need more persistent state remove the queryKey: false
-const appHistory = useRouterHistory(createHashHistory)({ queryKey: false})
-
-const auth = new AuthService(config.clientId, config.clientDomain);
 
 // onEnter callback to validate authentication in private routes
 const requireAuth = (nextState, replace) => {
-  if (!auth.loggedIn()) {
+  if (!AuthService.loggedIn()) {
     replace({ nextPathname: nextState.location.pathname }, '/login');
   }
 };
 
 ReactDOM.render(
-  <Router history={appHistory}>
-    <Route path="/" component={Layout} auth={auth}>
+  <Router history={AppHistory}>
+    <Route path="/" component={Layout}>
       <IndexRoute apiBaseURL="https://dev.calligre.com/api" component={Home} onEnter={requireAuth}></IndexRoute>
       <Route path="newsfeed" component={NewsFeed} onEnter={requireAuth}></Route>
       <Route path="people" component={People} onEnter={requireAuth}></Route>
@@ -42,8 +38,7 @@ ReactDOM.render(
       <Route path="events/:eventId" component={EventPage} onEnter={requireAuth}></Route>
       <Route path="profile" component={Profile} onEnter={requireAuth}></Route>
       <Route path="info" apiBaseURL="https://dev.calligre.com/api" component={Info} onEnter={requireAuth}></Route>
-      <Route path="login" component={Login} auth={auth}></Route>
-      <Route path="access_token=:token" component={Login} auth={auth}/>
+      <Route path="login" component={Login}></Route>
     </Route>
   </Router>,
 app);
