+++ src/js/organizer.client.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { Router, Route, IndexRoute, hashHistory, IndexLink } from "react-router";
+import { Router, Route, useRouterHistory, IndexRoute, IndexLink } from "react-router";
 
 require('!style!css!sass!../sass/main.scss');
 
@@ -14,20 +14,24 @@ import Profile from "./pages/Profile";
 import Info from "./pages/Info";
 import Login from "./pages/Login";
 
+import AppHistory from './util/AppHistory';
 import AuthService from './util/AuthService';
+import * as config from './auth0.config.js';
 
 const app = document.getElementById('app');
 
+const auth = new AuthService(config.clientId, config.clientDomain);
+
 // onEnter callback to validate authentication in private routes
 const requireAuth = (nextState, replace) => {
-  if (!AuthService.loggedIn()) {
+  if (!auth.loggedIn()) {
     replace({ nextPathname: nextState.location.pathname }, '/login');
   }
 };
 
 ReactDOM.render(
-  <Router history={hashHistory}>
-    <Route path="/" component={Layout}>
+  <Router history={AppHistory}>
+    <Route path="/" component={Layout} auth={auth}>
       <IndexRoute apiBaseURL="https://dev.calligre.com/api" component={Home} onEnter={requireAuth}></IndexRoute>
       <Route path="newsfeed" component={NewsFeed} onEnter={requireAuth}></Route>
       <Route path="people" component={People} onEnter={requireAuth}></Route>
@@ -36,8 +40,7 @@ ReactDOM.render(
       <Route path="events/:eventId" component={EventPage} onEnter={requireAuth}></Route>
       <Route path="profile" component={Profile} onEnter={requireAuth}></Route>
       <Route path="info" apiBaseURL="https://dev.calligre.com/api" component={Info} onEnter={requireAuth}></Route>
-      <Route path="login" component={Login}></Route>
-      <Route path="access_token=:token" component={Login}/>
+      <Route path="login" component={Login} auth={auth}></Route>
     </Route>
   </Router>,
 app);
